+++ lynks/helpers.py
@@ -10,3 +10,13 @@ def load_config(config_path: Path) -> dict:
         config = toml.load(f)
 
     return config
+
+
+def create_pipeline(list_functions):
+    def pipeline(*input):
+        res = input
+        for function in list_functions:
+            res = function(*res)
+        return res
+
+    return pipeline
